fix: extract BookingAuditErrorCode to shared file to avoid server-side imports in client component
