+++ packages/features/booking-audit/lib/BookingAuditErrorCode.ts
@@ -1,0 +1,7 @@
+export enum BookingAuditErrorCode {
+  ORGANIZATION_ID_REQUIRED = "ORGANIZATION_ID_REQUIRED",
+  BOOKING_NOT_FOUND_OR_PERMISSION_DENIED = "BOOKING_NOT_FOUND_OR_PERMISSION_DENIED",
+  BOOKING_HAS_NO_OWNER = "BOOKING_HAS_NO_OWNER",
+  OWNER_NOT_IN_ORGANIZATION = "OWNER_NOT_IN_ORGANIZATION",
+  ORG_MEMBER_PERMISSION_DENIED = "ORG_MEMBER_PERMISSION_DENIED",
+}
